add ContactForm component
